+++ R/defunct.R
@@ -5,6 +5,6 @@
 #' @details `docx_reference_img()` is not usefull anymore.
 #' @param ... unused arguments
 #' @keywords internal
-docx_reference_img <- function(...){
+docx_reference_img <- function(...) {
   .Defunct(msg = "The `docx_reference_img()` function is no longer useful.")
 }